+++ MonsterFitness/ProfileViewController.swift
@@ -102,7 +102,7 @@ class ProfileViewController: UIViewController {
         view.addSubview(userView.name)
         userView.name.translatesAutoresizingMaskIntoConstraints = false
         userView.name.leftAnchor.constraint(equalTo: view.leftAnchor, constant: 30).isActive = true
-        userView.name.topAnchor.constraint(equalTo: view.topAnchor, constant: 150).isActive = true
+        userView.name.topAnchor.constraint(equalTo: view.topAnchor, constant: 100).isActive = true
     }
     
     func settingsUserInformation() {
@@ -192,4 +192,31 @@ class ProfileViewController: UIViewController {
         }
         
     }
+}
+
+struct CONFIG {
+    static let backgroundColor: UIColor = UIColor(red: 13 / 255, green: 13 / 255, blue: 13 / 255, alpha: 1)
+    static let deviderColor: UIColor = UIColor(red: 40 / 255, green: 40 / 255, blue: 40 / 255, alpha: 1)
+    static let buttonBackgroudColor: UIColor = UIColor(red: 30 / 255,
+                                                       green: 30 / 255,
+                                                       blue: 30 / 255, alpha: 1)
+    static let buttonTextColor: UIColor = .white
+    static let buttonBorderColor: CGColor = CGColor(gray: 0.3, alpha: 1)
+    static let buttonBorderWidth: CGFloat = 1
+    static let buttonCornerRadius: CGFloat = 16
+    static let colorInactiveButton: CGFloat = 16
+
+    static let spacingBetweenElementsInPicker: CGFloat = 5
+    static let pikerButtonsShadowOpacity: Float = 0.6
+    static let pikerButtonsColor: CGColor = CGColor(red: 137 / 255,
+                                                    green: 143 / 255,
+                                                    blue: 85 / 255, alpha: 1)
+    static let pikerButtonTextColorInactive: UIColor = UIColor(red: 90 / 255,
+                                                    green: 90 / 255,
+                                                    blue: 90 / 255, alpha: 1)
+
+    static let segmentSelectedColor: UIColor = .systemGreen
+
+    static let searchFieldTextColor: UIColor = .white
+
 }
